Refactor log to its own sub dir
