+++ contrib/mesos/cmd/k8sm-controller-manager/main.go
@@ -23,6 +23,7 @@ import (
 
 	"k8s.io/kubernetes/pkg/healthz"
 	"k8s.io/kubernetes/pkg/util"
+	"k8s.io/kubernetes/pkg/util/log"
 	"k8s.io/kubernetes/pkg/version/verflag"
 
 	"k8s.io/kubernetes/contrib/mesos/pkg/controllermanager"
@@ -41,8 +42,8 @@ func main() {
 	s.AddFlags(pflag.CommandLine)
 
 	util.InitFlags()
-	util.InitLogs()
-	defer util.FlushLogs()
+	log.InitLogs()
+	defer log.FlushLogs()
 
 	verflag.PrintAndExitIfRequested()
