+++ contrib/mesos/cmd/k8sm-scheduler/main.go
@@ -25,6 +25,7 @@ import (
 	"k8s.io/kubernetes/contrib/mesos/pkg/hyperkube"
 	"k8s.io/kubernetes/contrib/mesos/pkg/scheduler/service"
 	"k8s.io/kubernetes/pkg/util"
+	"k8s.io/kubernetes/pkg/util/log"
 	"k8s.io/kubernetes/pkg/version/verflag"
 )
 
@@ -34,8 +35,8 @@ func main() {
 	s.AddStandaloneFlags(pflag.CommandLine)
 
 	util.InitFlags()
-	util.InitLogs()
-	defer util.FlushLogs()
+	log.InitLogs()
+	defer log.FlushLogs()
 
 	verflag.PrintAndExitIfRequested()
